Python parser complete. CSV's added.
Python parser is in working condition, but cleanup, optimization, and
generalization is required. cbr_summer.csv is the static CBR data given
to the WPI team the summer of 2016.  model_data.py.csv is the result of
the Python parser. parser.py will export to model_data.csv when it’s
adequately tested.